web_interface/requirements.txt
flask
flask-socketio
pyserial
requests
eventlet
faster-whisper
nvidia-cublas-cu12
nvidia-cudnn-cu12
nvidia-cuda-runtime-cu12